Make style analyser work
This commit contains several things. It is hard to split them into several because they are not going to work without each other. 
The best thing: this commit makes style analyzer produce meaningful comments. 

Changes list: 
1. New feature matrix. I add only several features and only for left tokens: one-hot-encoded (OHE) for reserved tokens, OHE for `y` (target) and 2 roles indicators: for `EXPRESSION` and `IDENTIFIER` roles. No right side, no parents, no self, no positions at all. Performance on reproduction task is `0.8484`. It is smaller than what we had, but comments are much better. You can check out usage example here: https://github.com/zurk/js_test_lookout/pull/3 
2. `humanize()` for `Rules` and `RuleAttribute` classes. Now we can convert unreadable rules into something useful:
```
    left_1_reserved in |(˙,˙;˙None˙{˙}|
    left_3_role in |expression|
    left_5_role in |expression|
```
`˙` is just separator. This part is connected to `1.` because of the special feature name structure. I think it can be avoided in the future. But now it is hard to do.
3. Output comment was improved a lot. I include all info that can help us better understand what is going on.
Signed-off-by: konstantin <kslavnov@gmail.com>